Fix mock services during the fixture loading
