+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 - Fix some documentation issue again.
 - Handles errors for the AbstractCommand.
 - Fix CommandTestCase bugs related to the usage of Helpers withing the command.
+- Fix not mocked services during the fixture loading.
+
+
+#### New features
+
 - Add handy functions for the fixture creation.
 - Add the `DATE_FORMAT` function to the SQLite language that behave like MySQL function.
